[🪟 Windows] Started Window API rework
