+++ MineSharp.Bot/Modules/Windows/WindowsModule.cs
@@ -61,19 +61,16 @@ namespace MineSharp.Bot.Modules.Windows
         public WindowsModule(MinecraftBot bot) : base(bot)
         {
             this.inventoryLoadedTsc = new TaskCompletionSource();
-            this.MainInventory = new Window(new WindowInfo((Identifier)"", "", 4 * 9, true));
-            this.MainInventory.WindowSlotUpdated += this.MainInventory_SlotUpdated;
+            this.MainInventory = new Window(0, 0, "MainInventory", 4 * 9, null, this.SynchronizeWindowClick);
+            this.MainInventory.OnSlotChanged += this.MainInventory_SlotUpdated;
         }
 
-        private Window MainInventory {
-            get;
-        }
-
+        private Window MainInventory { get; }
         public Window? Inventory { get; private set; }
 
 
         public byte SelectedHotbarIndex { get; private set; }
-        public Item? HeldItem => this.Inventory!.GetSlot((int)PlayerWindowSlots.HotbarStart + this.SelectedHotbarIndex).Item;
+        public Item? HeldItem => this.Inventory!.GetSlot((short)(PlayerWindowSlots.HotbarStart + this.SelectedHotbarIndex)).Item;
 
         public event BotWindowEvent? WindowOpened;
         public event BotItemEvent? HeldItemChanged;
@@ -122,15 +119,13 @@ namespace MineSharp.Bot.Modules.Windows
                 this.Logger.Warning("Tried to close window which is not open!");
                 return Task.CompletedTask;
             }
-
-            window.WindowClicked -= this.Window_Clicked;
             
             if (this.CurrentlyOpenedWindow == window)
             {
                 this.CurrentlyOpenedWindow = null;
             }
             
-            window.Close();
+            // TODO: window.Close();
             return this.Bot.Client.SendPacket(new PacketCloseWindow((byte)id));
         }
 
@@ -154,10 +149,13 @@ namespace MineSharp.Bot.Modules.Windows
                 throw new ArgumentException("Window with id " + id + " already opened");
             }
 
-            var window = new Window(id, windowInfo, playerInventory: windowInfo.ExcludeInventory ? null : this.MainInventory) {
-                IsSynchronized = true
-            };
-            window.WindowClicked += this.Window_Clicked;
+            var window = new Window(
+                (byte)id, 
+                -1, 
+                windowInfo.Title, 
+                windowInfo.UniqueSlots, 
+                windowInfo.ExcludeInventory ? null : this.MainInventory, 
+                this.SynchronizeWindowClick);
 
             this.OpenedWindows.Add(id, window);
             this.WindowOpened?.Invoke(this.Bot, window);
@@ -180,14 +178,20 @@ namespace MineSharp.Bot.Modules.Windows
             return window;
         }
 
-        private void Window_Clicked(Window window, WindowClick click)
+        private Task SynchronizeWindowClick(Window window, WindowClick click)
         {
             var windowClickPacket = new PacketWindowClick(
-                (byte)window.Id, window.StateId, click.Slot, (sbyte)click.Button, (sbyte)click.ClickMode, window.GetAllSlots().Select(x => new PacketWindowClick.ChangedSlotsElementContainer(x.SlotNumber, x.ToProtocolSlot())).ToArray(), window.SelectedSlot!.ToProtocolSlot());
-            this.Bot.Client.SendPacket(windowClickPacket).ContinueWith(x => window.SetSync());
+                window.WindowId, 
+                window.StateId,
+                click.Slot,
+                (sbyte)click.Button,
+                (sbyte)click.ClickMode,
+                window.GetAllSlots().Select(x => new PacketWindowClick.ChangedSlotsElementContainer(x.SlotNumber, x.ToProtocolSlot())).ToArray(), 
+                window.GetSelectedSlot().ToProtocolSlot());
+            return this.Bot.Client.SendPacket(windowClickPacket);
         }
 
-        private void MainInventory_SlotUpdated(Window window, int index)
+        private void MainInventory_SlotUpdated(Window window, short index)
         {
             if (index == 3 * 9 + 1 + this.SelectedHotbarIndex)
             {
@@ -216,18 +220,7 @@ namespace MineSharp.Bot.Modules.Windows
                 return Task.CompletedTask;
             }
             
-
-            var slot = packet.Item!.ToSlot();
-            slot.SlotNumber = packet.Slot!;
-
-            if (slot.SlotNumber == -1)
-            { // used to set selecteditem
-                window.SelectedSlot = slot;
-            } else
-            {
-                window.SetSlot(slot);
-            }
-
+            window.SetProtocolSlot(packet.Item, packet.Slot);
             window.StateId = packet.StateId!;
 
             return Task.CompletedTask;
@@ -237,7 +230,6 @@ namespace MineSharp.Bot.Modules.Windows
         private PacketWindowItems? cachedWindowItemsPacket;
         private Task handleWindowItems(PacketWindowItems packet)
         {
-
             if (!this.OpenedWindows.TryGetValue(packet.WindowId!, out var window))
             {
                 this.Logger.Warning($"Received {packet.GetType().Name} for windowId={packet.WindowId!}, but its not opened");
@@ -248,15 +240,10 @@ namespace MineSharp.Bot.Modules.Windows
                 return Task.CompletedTask;
             }
 
-            window.UpdateSlots(packet.Items!.Select((x, i) =>
-            {
-                var slot = x.ToSlot();
-                slot.SlotNumber = (short)i;
-                return slot;
-            }).ToArray());
+            window.SetProtocolSlots(packet.Items);
             window.StateId = packet.StateId!;
 
-            if (window.Id == 0 && !this.inventoryLoadedTsc.Task.IsCompleted)
+            if (window.WindowId == 0 && !this.inventoryLoadedTsc.Task.IsCompleted)
             {
                 this.inventoryLoadedTsc.SetResult();
             }
